Add utils test to check revoke func
