+++ test/conftest.py
@@ -1,10 +1,16 @@
+
+from cryptography.hazmat.primitives.asymmetric import rsa
+
+
+from cryptography import x509
+from cryptography.x509 import NameOID
+
 import datetime
 
 import pytest
-from cryptography import x509
 from cryptography.hazmat.primitives import hashes, serialization
-from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography.x509 import NameOID
+from cryptography.hazmat.primitives._serialization import Encoding
+from cryptography.x509 import ocsp
 
 
 TEST_DISTRIBUTION_POINT_URL = "test_url"
@@ -107,3 +113,51 @@ def _create_cert(key_pair, add_crl_extension=True, add_aia_extension=True):
 @pytest.fixture()
 def cert_pem_string(cert):
     return cert.public_bytes(serialization.Encoding.PEM).decode()
+
+
+
+def _create_mocked_ocsp_response(
+    cert, key_pair, status=ocsp.OCSPCertStatus.GOOD, revocation_time=None
+):
+    builder = ocsp.OCSPResponseBuilder()
+    builder = builder.add_response(
+        cert=cert,
+        issuer=cert,
+        algorithm=hashes.SHA256(),
+        cert_status=status,
+        this_update=datetime.datetime.now(),
+        next_update=datetime.datetime.now(),
+        revocation_time=revocation_time,
+        revocation_reason=None,
+    ).responder_id(ocsp.OCSPResponderEncoding.HASH, cert)
+    return builder.sign(key_pair, hashes.SHA256()).public_bytes(Encoding.DER)
+
+
+def _create_crl(keypair, revoked_serials):
+    one_day = datetime.timedelta(days=1)
+    crl = x509.CertificateRevocationListBuilder()
+    crl = crl.issuer_name(
+        x509.Name(
+            [
+                x509.NameAttribute(NameOID.COMMON_NAME, "cryptography.io CA"),
+            ]
+        )
+    )
+    crl = crl.last_update(datetime.datetime.today())
+    crl = crl.next_update(datetime.datetime.today() + one_day)
+
+    for serial in revoked_serials:
+        next_revoked_cert = (
+            x509.RevokedCertificateBuilder()
+            .serial_number(
+                serial,
+            )
+            .revocation_date(
+                datetime.datetime.today(),
+            )
+            .build()
+        )
+
+        crl = crl.add_revoked_certificate(next_revoked_cert)
+
+    return crl.sign(private_key=keypair, algorithm=hashes.SHA256())
